core_services/model_loading/requirements.txt
sentence-transformers
transformers
grpcio
grpcio-tools
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-grpc
prometheus_client
pytest
pytest-asyncio
hf_xet
huggingface_hub[hf_xet]